requirements.txt
hyperliquid-python-sdk
discord
discord.py
ccxt
requests
pandas
numpy
pandas_ta
memory-profiler
flask
